% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareaPD.R
\name{compareaPD}
\alias{compareaPD}
\title{compareaPD}
\usage{
compareaPD(x, phy, area)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree of class phylo}

\item{area}{name of selected area to measure complementary phylogenetic diversity
of the other areas in the set}
}
\value{

}
\description{
compareaPD
}
