% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemismPD.R
\name{endemismPD}
\alias{endemismPD}
\title{endemismPD}
\usage{
endemismPD(x, phy, plot = T)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree(s) of class phylo or multiPhylo}

\item{plot}{logical value indicating whether to plot the input tree. Defaults to
plot = TRUE unless more than one input tree is provided.}
}
\value{
An object of class “vector” or "matrix" containing the endemism phylogenetic
diversity values for each area given all input trees.
A plot showing the branches used for calculating the endemism PD of
a given area (in red). The endemism PD value is given in the title of each
plot.
}
\description{
endemismPD
}
\details{
endemismPD calculates the endemism phylogenetic diversity of an area given the
other areas in the set. When there is a single input tree, the function plots
the branches used for calculating the endemism PD of a given area (in red).
Currently, the function is implemented to deal with five areas only.
}
\examples{
#endemism PD calculation for a set of areas given a phylogeny
data("homodata")
data("homotree")
homoPD<-endemismPD(homodata[1:5,],homotree)
}
