% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVennPD.R
\name{makeVennPD}
\alias{makeVennPD}
\title{makeVennPD}
\usage{
makeVennPD(x)
}
\arguments{
\item{x}{list of vectors containing phylogenetic diversity values within
and between areas produced by multiareaPD}
}
\value{

}
\description{
makeVennPD
}
