% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiareaPD.R
\name{multiareaPD}
\alias{multiareaPD}
\title{multiareaPD}
\usage{
multiareaPD(x, phy)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree of class phylo}
}
\value{
An object of class “vector” containing the phylogenetic diversity
values for each area and for all combinations of areas.
}
\description{
multiareaPD
}
\details{
multiareaPD uses the function picante::pd to calculate phylogenetic diversity
for a set of areas. It plots the input tree showing the branches used
for calculating the PD of a given area.It also plots a Venn's
diagram based on shared patterns of PD among areas. Currently,
the function is implemented to deal with five areas only.
}
\examples{
#PD calculation for a set of areas and a given phylogeny
data("homodata")
data("homotree")
homoPD<-multiareaPD(homodata,homotree)

}
