% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareaPD.R
\name{compareaPD}
\alias{compareaPD}
\title{compareaPD}
\usage{
compareaPD(x, phy, area)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree of class phylo}

\item{area}{name of selected area to measure complementary phylogenetic diversity
of the other areas in the set}
}
\value{
A plot showing the branches used for calculating the complementary PD of
a given area (in red) and the branches representing the PD of the selected
area (in green). The complementary PD value is given in the title of each
plot.
}
\description{
compareaPD
}
\details{
compareaPD calculates the complementary phylogenetic diversity of an area
given another selected area from the set. It also plots the input tree
showing the branches used for calculating the complementary PD of a given
area (in red) and the branches representing the PD of the selected area.
}
\examples{
#complementary PD calculation for a set of areas given a phylogeny and
#a selected area
data("homodata")
data("homotree")
homoPD<-compareaPD(homodata,homotree,LCL)
}
