% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiareaPD.R
\name{multiareaPD}
\alias{multiareaPD}
\title{multiareaPD}
\usage{
multiareaPD(x, phy)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree of class phylo}
}
\value{

}
\description{
multiareaPD
}
