% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiareaPD.R
\name{multiareaPD}
\alias{multiareaPD}
\title{multiareaPD}
\usage{
multiareaPD(x, phy, plot = T)
}
\arguments{
\item{x}{dataframe containing the assignment of samples (rows) to areas (columns)}

\item{phy}{phylogenetic tree(s) of class phylo or multiPhylo}

\item{plot}{logical value indicating whether to plot the input tree. Defaults to
plot = TRUE unless more than one input tree is provided.}
}
\value{
An object of class “vector” or "matrix" containing the phylogenetic diversity
values for each area and for all combinations of areas given all input trees.
}
\description{
multiareaPD
}
\details{
multiareaPD uses the function picante::pd to calculate phylogenetic diversity
for a set of areas. When there is a single input tree, the function plots the
branches used for calculating the PD of a given area in red color. It also
plots a Venn's diagram based on shared patterns of PD among areas. Currently,
the function is implemented to deal with five areas only.
}
\examples{
#PD calculation for a set of areas and a given phylogeny
data("homodata")
data("homotree")
homoPD<-multiareaPD(homodata,homotree)

}
